[wip, need to add tests] move generic code out of uihandler
